Use file basename instead of fullname in sonar
Some sonar plugin do no include git path as the source file path (for
example, sonar-visual-studio plugin generates filenames that are
relative to the module csproj file).

Using the file basename and a recursive match allows Sonar to match
modified files with their indexed names in Sonar.

Although this can generate useless analysis (as some files that are not
included in a review may be analysed), this has a limited additional
cost considering that few files have the same basename inside a
repository.

Change-Id: Ib74289e201098d15c618791f47b2dd657c3a1744